move declaration samples
Move all samples from the /samples/snippets folder under the current folder.
