+++ docs/csharp/language-reference/statements/declarations.md
@@ -41,31 +41,11 @@ In pattern matching, the `var` keyword is used in a [`var` pattern](../operators
 
 The following example shows two query expressions. In the first expression, the use of `var` is permitted but is not required, because the type of the query result can be stated explicitly as an `IEnumerable<string>`. However, in the second expression, `var` allows the result to be a collection of anonymous types, and the name of that type is not accessible except to the compiler itself. Use of `var` eliminates the requirement to create a new class for the result. Note that in Example #2, the `foreach` iteration variable `item` must also be implicitly typed.
 
-[!code-csharp[csrefKeywordsTypes#18](~/samples/snippets/csharp/VS_Snippets_VBCSharp/csrefKeywordsTypes/CS/keywordsTypes.cs#18)]
-
-## Ref returns and ref locals
-
-Starting with C# 7.0, C# supports reference return values (ref returns). A reference return value allows a method to return a reference to a variable, rather than a value, back to a caller. The caller can then choose to treat the returned variable as if it were returned by value or by reference. The caller can create a new variable that is itself a reference to the returned value, called a ref local.
-
-Most developers are familiar with passing an argument to a called method *by reference*. A called method's argument list includes a variable passed by reference. Any changes made to its value by the called method are observed by the caller. A *reference return value* means that a method returns a *reference* (or an alias) to some variable. That variable's scope must include the method. That variable's lifetime must extend beyond the return of the method. Modifications to the method's return value by the caller are made to the variable that is returned by the method.
-
-Declaring that a method returns a *reference return value* indicates that the method returns an alias to a variable. The design intent is often that the calling code should have access to that variable through the alias, including to modify it. It follows that methods returning by reference can't have the return type `void`.
-
-There are some restrictions on the expression that a method can return as a reference return value. Restrictions include:
-
-- The return value must have a lifetime that extends beyond the execution of the method. In other words, it cannot be a local variable in the method that returns it. It can be an instance or static field of a class, or it can be an argument passed to the method. Attempting to return a local variable generates compiler error CS8168, "Cannot return local 'obj' by reference because it is not a ref local."
-
-- The return value cannot be the literal `null`. Returning `null` generates compiler error CS8156, "An expression cannot be used in this context because it may not be returned by reference."
-
-   A method with a ref return can return an alias to a variable whose value is currently the null (uninstantiated) value or a [nullable value type](../../language-reference/builtin-types/nullable-value-types.md) for a value type.
-
-- The return value cannot be a constant, an enumeration member, the by-value return value from a property, or a method of a `class` or `struct`. Violating this rule generates compiler error CS8156, "An expression cannot be used in this context because it may not be returned by reference."
-
-In addition, reference return values are not allowed on async methods. An asynchronous method may return before it has finished execution, while its return value is still unknown.
+:::code language="csharp" source="./snippets/declarations/ImplicitlyTyped.cs" id="VarExample":::
 
 ## Ref locals
 
-Assume the `GetContactInformation` method is declared as a ref return:
+Assume the `GetContactInformation` method is declared as a [ref return](jump-statements.md#ref-returns):
 
 ```csharp
 public ref Person GetContactInformation(string fname, string lname)
@@ -108,27 +88,26 @@ refLocal = ref anotherVeryLargeStruct; // reassigned, refLocal refers to differe
 
 The following example defines a `NumberStore` class that stores an array of integer values. The `FindNumber` method returns by reference the first number that is greater than or equal to the number passed as an argument. If no number is greater than or equal to the argument, the method returns the number in index 0.
 
-[!code-csharp[ref-returns](../../../../samples/snippets/csharp/programming-guide/ref-returns/NumberStore.cs#1)]
+:::code language="csharp" source="./snippets/declarations/NumberStore.cs" id="Snippet1":::
 
 The following example calls the `NumberStore.FindNumber` method to retrieve the first value that is greater than or equal to 16. The caller then doubles the value returned by the method. The output from the example shows the change reflected in the value of the array elements of the `NumberStore` instance.
 
-[!code-csharp[ref-returns](../../../../samples/snippets/csharp/programming-guide/ref-returns/NumberStore.cs#2)]
+:::code language="csharp" source="./snippets/declarations/NumberStore.cs" id="Snippet2":::
 
 Without support for reference return values, such an operation is performed by returning the index of the array element along with its value. The caller can then use this index to modify the value in a separate method call. However, the caller can also modify the index to access and possibly modify other array values.  
 
 The following example shows how the `FindNumber` method could be rewritten after
 C# 7.3 to use ref local reassignment:
 
-[!code-csharp[ref-returns](../../../../samples/snippets/csharp/programming-guide/ref-returns/NumberStoreUpdated.cs#1)]
+:::code language="csharp" source="./snippets/declarations/NumberStoreUpdated.cs" id="Snippet1":::
 
 This second version is more efficient with longer sequences in scenarios where the number sought is
 closer to the end of the array, as the array is iterated from end towards the beginning, causing fewer items to be examined.
 
 ## See also
 
-- [ref keyword](../../language-reference/keywords/ref.md)
+- [ref keyword](../keywords/ref.md)
 - [Write safe efficient code](../../write-safe-efficient-code.md)
 - ['var' preferences (style rules IDE0007 and IDE0008)](../../../fundamentals/code-analysis/style-rules/ide0007-ide0008.md)
 - [C# reference](../index.md)
-- [Implicitly typed local variables](../../programming-guide/classes-and-structs/implicitly-typed-local-variables.md)
 - [Type relationships in LINQ query operations](../../programming-guide/concepts/linq/type-relationships-in-linq-query-operations.md)
