+++ docs/csharp/language-reference/statements/snippets/declarations/NumberStore.cs
