+++ docs/csharp/language-reference/statements/snippets/declarations/NumberStoreUpdated.cs
