+++ docs/csharp/language-reference/statements/snippets/declarations/Program.cs
@@ -1,0 +1,9 @@
+﻿using RefReturns;
+using RefReturnsVersionTwo;
+using ImplicitTypes;
+
+FirstExample.Tests();
+Console.WriteLine();
+SecondExample.Tests();
+Console.WriteLine();
+VarExample.ImplicitlyTyped();
