--- samples/snippets/csharp/programming-guide/ref-returns/Program.cs
@@ -1,16 +1,0 @@
-﻿namespace UseRefReturns
-{
-    using static System.Console;
-    using RefReturns;
-    using RefReturnsVersionTwo;
-
-    public class Example
-    {
-       static void Main(string[] args)
-       {
-            FirstExample.Tests();
-            WriteLine();
-            SecondExample.Tests();
-       }
-    }
-}
