--- samples/snippets/csharp/programming-guide/ref-returns/ref-returns.csproj
@@ -1,10 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <LangVersion>7.3</LangVersion>
-    <RootNamespace>RefReturns</RootNamespace>
-  </PropertyGroup>
-
-</Project>
